+++ src/components/TeamPage/__subComponents/HistoryCell.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { TableCell } from '@mui/material';
+
+interface HistoryCellProps {
+    backgroundColor?: string;
+    isBold?: boolean;
+}
+
+export const HistoryCell: React.FC<HistoryCellProps> = ({
+    children,
+    backgroundColor,
+    isBold,
+}) => (
+    <TableCell
+        sx={{
+            border: '1px solid gray',
+            backgroundColor,
+            fontWeight: isBold ? 'bold' : 'auto',
+        }}
+        align="center"
+    >
+        {children}
+    </TableCell>
+);