+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Features:
 * Tournament places and points are now bigger.
 * Teams with incomplete roster display with red names now.
 * Players are now sorted by positions, teams by fame.
+* Stores and displays team main roster history.
 
 Fixes: